requirements.txt
pandas==1.5.3
streamlit==1.21.0
biopython==1.81